Remove restriction that TaskPreserving may not be TaskEnabled
As interceptors run before decorators, the TaskScopeEnabledInterceptor may run before the preserving decorator, starting a new task. However the decorator enters the previous (and preserved) task again, leaving the interceptor started task useless. This is useless but not harmful and the deployment time checks are too hard to implement, so the restriction is removed